pyproject.toml
[tool.poetry]
name = "socially"
version = "0.1.0"
description = "Social network website created in django"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.13"
psycopg2-binary = "^2.9.3"
django-environ = "^0.5.0"
Pillow = "^9.0.1"
django-crispy-forms = "^1.12.0"
django-allauth = "^0.45.0"
crispy-bootstrap5 = "^0.5"
django-compressor = "^2.4.1"
django-libsass = "^0.9"
coverage = "^5.5"
factory-boy = "^3.2.0"
Faker = "^8.14.1"
django-htmx = "^1.6.0"
djangorestframework = "^3.13.1"
dj-rest-auth = { extras = ["with_social"], version = "^2.2.4" }
djangorestframework-simplejwt = "^5.1.0"
django-cors-headers = "^3.11.0"

[tool.poetry.dev-dependencies]
black = { version = "^21.8b0", allow-prereleases = true }
pytest = "^7.1.2"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''
